requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bcrypt==3.2.0
bleach==3.2.1
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dnspython==2.0.0
doit==0.33.1
email-validator==1.1.1
entrypoints==0.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
ipykernel==5.3.4
ipympl==0.5.8
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-server==1.0.4
jupyterlab-pygments==0.1.2
jupytext==1.6.0
kiwisolver==1.2.0
lxml==4.5.2
markdown-it-py==0.5.5
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
mysql-connector==2.2.9
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
numexpr==2.7.1
numpy==1.19.2
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
pyinotify==0.9.6
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
PyYAML==5.3.1
pyzmq==19.0.2
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
spike-py==0.99.19
SQLAlchemy==1.3.19
tables==3.6.1
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
voila==0.2.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
WTForms==2.3.3